Data context fixed, template updated
Need to use injury.parameter to pull that from the template helper,
which was done on a previous version of the app.
updated fixtures and template to show more injury data like date and
notes.
next need to created edit, upsert functions and also attach the rehab
schema to select interval running exercises for each injury through a
date picker / auto form